+++ R/cv_varsel.R
@@ -1462,7 +1462,11 @@ get_kfold <- function(refmodel, K, cvfits, verbose) {
 #'   [init_refmodel()]) or an object that can be passed to argument `object` of
 #'   [get_refmodel()].
 #' @param K Number of folds. Must be at least 2 and not exceed the number of
-#'   observations.
+#'   observations. Ignored if `folds` is not `NULL`.
+#' @param folds Either `NULL` for determining the CV folds automatically via
+#'   [cv_folds()] (using argument `K`) or a numeric (in fact, integer) vector
+#'   giving the fold index for each observation. In the latter case, argument
+#'   `K` is ignored.
 #' @param seed Pseudorandom number generation (PRNG) seed by which the same
 #'   results can be obtained again if needed. Passed to argument `seed` of
 #'   [set.seed()], but can also be `NA` to not call [set.seed()] at all. If not
@@ -1513,6 +1517,23 @@ get_kfold <- function(refmodel, K, cvfits, verbose) {
 #'                      cvfits = cv_fits, nterms_max = 3, nclusters = 5,
 #'                      nclusters_pred = 10, seed = 5555)
 #'
+#' # Stratified K-fold CV is straightforward:
+#' n_strat <- 3L
+#' set.seed(692)
+#' # Some example strata:
+#' strat_fac <- sample(paste0("lvl", seq_len(n_strat)), size = nrow(dat_gauss),
+#'                     replace = TRUE,
+#'                     prob = diff(c(0, pnorm(seq_len(n_strat - 1L) - 0.5), 1)))
+#' table(strat_fac)
+#' # Use loo::kfold_split_stratified() to create the folds vector:
+#' folds_strat <- loo::kfold_split_stratified(K = 2, x = strat_fac)
+#' table(folds_strat, strat_fac)
+#' # Call run_cvfun(), but this time with argument `folds` instead of `K` (here,
+#' # specifying argument `seed` would not be necessary because of the set.seed()
+#' # call above, but we specify it nonetheless for the sake of generality):
+#' cv_fits_strat <- run_cvfun(ref, folds = folds_strat, seed = 391)
+#' # Now use `cv_fits_strat` analogously to `cv_fits` from above.
+#'
 #' @export
 run_cvfun <- function(object, ...) {
   UseMethod("run_cvfun")
@@ -1529,7 +1550,7 @@ run_cvfun.default <- function(object, ...) {
 #' @export
 run_cvfun.refmodel <- function(object,
                                K = if (!inherits(object, "datafit")) 5 else 10,
-                               seed = NA, ...) {
+                               folds = NULL, seed = NA, ...) {
   if (exists(".Random.seed", envir = .GlobalEnv)) {
     rng_state_old <- get(".Random.seed", envir = .GlobalEnv)
   }
@@ -1544,7 +1565,9 @@ run_cvfun.refmodel <- function(object,
   refmodel <- object
   stopifnot(!is.null(refmodel$cvfun))
 
-  folds <- cv_folds(refmodel$nobs, K = K)
+  if (is.null(folds)) {
+    folds <- cv_folds(refmodel$nobs, K = K)
+  }
   if (getOption("projpred.warn_kfold_refits", TRUE)) {
     cvfits <- refmodel$cvfun(folds)
   } else {